nuevas cositas

+++ session/session.go
@@ -15,7 +15,7 @@ import (
 type Id string
 
 type User struct {
-	Id   int
+	Id   int64
 	Name string
 	Role core.Role
 }
